LU-1436 debug: return 64 bits for libcfs_log_return
in libcfs_log_return, the return type (long) is 32 bits for i686,
which will truncate the 64 bits value passed through RETURN() macro.

+++ libcfs/include/libcfs/libcfs_debug.h
@@ -299,14 +299,14 @@ do {                                                                    \
  */
 #if defined(__GNUC__)
 
-long libcfs_log_return(struct libcfs_debug_msg_data *, long rc);
+long long libcfs_log_return(struct libcfs_debug_msg_data *, long long rc);
 #define RETURN(rc)                                                      \
 do {                                                                    \
         EXIT_NESTING;                                                   \
         if (cfs_cdebug_show(D_TRACE, DEBUG_SUBSYSTEM)) {                \
                 LIBCFS_DEBUG_MSG_DATA_DECL(msgdata, D_TRACE, NULL);     \
                 return (typeof(rc))libcfs_log_return(&msgdata,          \
-                                                     (long)(rc));       \
+                                                     (long long)(rc));  \
         }                                                               \
                                                                         \
         return (rc);                                                    \
